+++ networkx/algorithms/spanner.py
